teacher some policy and permission is edit and composer update

+++ modules/Hadihosseini88/Course/Resources/Views/details.blade.php
@@ -13,13 +13,14 @@
                     <p class="mlg-15">{{ $course->title }}</p>
                     <a class="color-2b4a83" href="{{ route('lessons.create', $course->id) }}">آپلود جلسه جدید</a>
                 </div>
-                <div class="d-flex item-center flex-wrap margin-bottom-15 operations__btns">
-                    <button class="btn all-confirm-btn" onclick="acceptAllLessons('{{ route('lessons.acceptAll', $course->id) }}')">تایید همه جلسات</button>
-                    <button class="btn confirm-btn" onclick="acceptMultiple('{{ route('lessons.acceptMultiple', $course->id) }}')">تایید جلسات</button>
-                    <button class="btn reject-btn" onclick="rejectMultiple('{{ route('lessons.rejectMultiple', $course->id) }}')">رد جلسات</button>
-                    <button class="btn delete-btn" onclick="deleteMultiple('{{ route('lessons.destroyMultiple', $course->id) }}')">حذف جلسات</button>
-
-                </div>
+                @can(\Hadihosseini88\RolePermissions\Models\Permission::PERMISSION_MANAGE_COURSES)
+                    <div class="d-flex item-center flex-wrap margin-bottom-15 operations__btns">
+                        <button class="btn all-confirm-btn" onclick="acceptAllLessons('{{ route('lessons.acceptAll', $course->id) }}')">تایید همه جلسات</button>
+                        <button class="btn confirm-btn" onclick="acceptMultiple('{{ route('lessons.acceptMultiple', $course->id) }}')">تایید جلسات</button>
+                        <button class="btn reject-btn" onclick="rejectMultiple('{{ route('lessons.rejectMultiple', $course->id) }}')">رد جلسات</button>
+                        <button class="btn delete-btn" onclick="deleteMultiple('{{ route('lessons.destroyMultiple', $course->id) }}')">حذف جلسات</button>
+                    </div>
+                @endcan
                 <div class="table__box">
                     <table class="table">
                         <thead role="rowgroup">
@@ -57,17 +58,19 @@
                                 <td>{{ $lesson->is_free ? 'همه' : 'شرکت کنندگان' }}</td>
                                 <td class="status {{ $lesson->getStatusCssClass() }}">@lang($lesson->status)</td>
                                 <td>
-                                    <a onclick="deleteItem(event,'{{route('lessons.destroy',[$course->id, $lesson->id])}}')" class="item-delete mlg-15" data-id="1" href="javascript: void(0)"
-                                       title="حذف"></a>
-                                    <a href="" onclick="updateConfirmationStatus(event,'{{route('lessons.reject', $lesson->id)}}','آیا از رد این آیتم اطمینان دارید؟','رد شده')" class="item-reject mlg-15" title="رد"></a>
-                                    @if($lesson->status == \Hadihosseini88\Course\Models\Lesson::STATUS_OPENED)
-                                        <a href="" onclick="updateConfirmationStatus(event,'{{route('lessons.lock', $lesson->id)}}','آیا از قفل این آیتم اطمینان دارید؟','قفل شده','status')"
-                                           class="item-lock mlg-15 text-error" title="قفل"></a>
-                                    @else
-                                        <a href="" onclick="updateConfirmationStatus(event,'{{route('lessons.unlock', $lesson->id)}}','آیا از باز این آیتم اطمینان دارید؟','باز','status')"
-                                           class="item-lock mlg-15 text-success" title="باز کردن"></a>
-                                    @endif
-                                    <a href="" onclick="updateConfirmationStatus(event,'{{route('lessons.accept', $lesson->id)}}','آیا از تایید این آیتم اطمینان دارید؟','تایید شده')" class="item-confirm mlg-15" title="تایید"></a>
+                                    @can(\Hadihosseini88\RolePermissions\Models\Permission::PERMISSION_MANAGE_COURSES)
+                                        <a onclick="deleteItem(event,'{{route('lessons.destroy',[$course->id, $lesson->id])}}')" class="item-delete mlg-15" data-id="1" href="javascript: void(0)"
+                                           title="حذف"></a>
+                                        <a href="" onclick="updateConfirmationStatus(event,'{{route('lessons.reject', $lesson->id)}}','آیا از رد این آیتم اطمینان دارید؟','رد شده')" class="item-reject mlg-15" title="رد"></a>
+                                        @if($lesson->status == \Hadihosseini88\Course\Models\Lesson::STATUS_OPENED)
+                                            <a href="" onclick="updateConfirmationStatus(event,'{{route('lessons.lock', $lesson->id)}}','آیا از قفل این آیتم اطمینان دارید؟','قفل شده','status')"
+                                               class="item-lock mlg-15 text-error" title="قفل"></a>
+                                        @else
+                                            <a href="" onclick="updateConfirmationStatus(event,'{{route('lessons.unlock', $lesson->id)}}','آیا از باز این آیتم اطمینان دارید؟','باز','status')"
+                                               class="item-lock mlg-15 text-success" title="باز کردن"></a>
+                                        @endif
+                                        <a href="" onclick="updateConfirmationStatus(event,'{{route('lessons.accept', $lesson->id)}}','آیا از تایید این آیتم اطمینان دارید؟','تایید شده')" class="item-confirm mlg-15" title="تایید"></a>
+                                    @endcan
                                     <a href="{{ route('lessons.edit',[$course->id, $lesson->id]) }}" class="item-edit " title="ویرایش"></a>
                                 </td>
                             </tr>
