+++ modules/Hadihosseini88/Course/Resources/Views/index.blade.php
@@ -84,18 +84,20 @@
                         <td>{{ $course->percent }}%</td>
                         <td class="status">@lang($course->status)</td>
                         <td>
-                            <a href="" onclick="deleteItem(event,'{{route('courses.destroy', $course->id)}}')"
-                               class="item-delete mlg-15" title="حذف"></a>
-                            <a href=""
-                               onclick="updateConfirmationStatus(event,'{{route('courses.reject', $course->id)}}','آیا از رد این آیتم اطمینان دارید؟','رد شده')"
-                               class="item-reject mlg-15" title="رد"></a>
-                            <a href=""
-                               onclick="updateConfirmationStatus(event,'{{route('courses.lock', $course->id)}}','آیا از قفل این آیتم اطمینان دارید؟','قفل شده','status')"
-                               class="item-lock mlg-15" title="قفل دوره"></a>
+                            @can(\Hadihosseini88\RolePermissions\Models\Permission::PERMISSION_MANAGE_COURSES)
+                                <a href="" onclick="deleteItem(event,'{{route('courses.destroy', $course->id)}}')"
+                                   class="item-delete mlg-15" title="حذف"></a>
+                                <a href=""
+                                   onclick="updateConfirmationStatus(event,'{{route('courses.lock', $course->id)}}','آیا از قفل این آیتم اطمینان دارید؟','قفل شده','status')"
+                                   class="item-lock mlg-15" title="قفل دوره"></a>
+                                <a href=""
+                                   onclick="updateConfirmationStatus(event,'{{route('courses.reject', $course->id)}}','آیا از رد این آیتم اطمینان دارید؟','رد شده')"
+                                   class="item-reject mlg-15" title="رد"></a>
+                                <a href=""
+                                   onclick="updateConfirmationStatus(event,'{{route('courses.accept', $course->id)}}','آیا از تایید این آیتم اطمینان دارید؟','تایید شده')"
+                                   class="item-confirm mlg-15" title="تایید"></a>
+                            @endcan
                             <a href="" target="_blank" class="item-eye mlg-15" title="مشاهده"></a>
-                            <a href=""
-                               onclick="updateConfirmationStatus(event,'{{route('courses.accept', $course->id)}}','آیا از تایید این آیتم اطمینان دارید؟','تایید شده')"
-                               class="item-confirm mlg-15" title="تایید"></a>
                             <a href="{{ route('courses.edit', $course->id) }}" class="item-edit " title="ویرایش"></a>
                         </td>
                     </tr>
